refactor: Comments 폴더를 CommentList 로 변경
